Added basic expression evaluation
The evaluation logic has some flaws:
- Can't handle parenthesis (see skipped tests)
- Needs spaces in the expressions

Before I fix those, I have to do a refactor:
The dependency graph of the classes is spiderweb-like.
I have to create interfaces for the core and
put them in a separate file.

But the rep 4 { f 100; l (360 / 4)} command draws a rectangle
which is cool