UI: Fix wrong fill function
Strange error: Couldn't fill with black color after a fresh start. Filling with black worked after filled with white or any other color.

The reason was that by default the canvas data was all-zero, which means that it was black with 100% transparency.

Originally I tried to handle when alpha is 0, but it was easier to draw a white rectangle. :)