CORE: Fix bad parenthesis handling
This is a great example where breaking the first point of SOLID (single
responsibility) architecturally causes a hidden bug.

The tokenizer did two tings:
1) Created a series of tokens from a string
2) Removed the parenthesis when they were just wrapped the whole
   expression

For example: From (1+2+3) it created a token 1+2+3
The reason behind this was that the parser creates separate tokens when
the string contains a space, but it ignores space when it is inside
parenthesis, so using the (1 + 2 + 3) format makes sense because of how
the tokenizer works. This was unnecessary, because (1 + 2 + 3) is a
valid expression, so the tokenizer should keep it.
Also it was implemented badly.