Refactoring the return and returnValue
The return flag and returnValue became independent.
The return flag @ CommandControl means that the command list execution
must be halted and it must throw back to the caller if the current
command list is not at command level.

The returnValue is independent from this.

So when a builtin command wants to give back some value, it must set the
returnValue, but not the return flag.