CORE: Fixed the usage of expressions stated with a valid variable
This was a strange bug, expressions like foo[0]+1 failed while 1+foo[0] was correct.
The issue was that the hasVariable checked if an expression has a structured addressing and if it is, then checked only the first part and handled the whole expression as a variable.

The point is that hasVariable and getVariable must receive only a variable, not a whole expression