UI: Add line numbers and refactor the Code editor
This was a major modification. I wanted to add line numbers to the
editor, but it was a bloated, badly organized code, so I split it to
follow S from SOLID.

The CodeEditor view now is very short, it only calls the relevant
functions. A separate hook (useSharedData) handles the SharedData.

Every component has its own responsibility and file.