Refactore and more cache -> 13.6 sec to 10.8 sec
I refactored the polish form evaluation, the original evaluatePolishForm was split into compile polish form, that did a some string checks and evaluated the expression into
- a pure compile step, that creates some instances
- a pure executing step, that just executes a series of object

Only the executing step depends on the memory, so everything else can be cached. (Not just the polish form.)
So skipping the compile step after a cache, the execution time reduced to its approx. 75%.

This compiled level is almost the same as LUA or Java byte code.